requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pgvector==0.3.3
python-dotenv==1.0.1
httpx==0.27.0
PyYAML==6.0.2
numpy==1.26.4
scikit-learn==1.5.1
